phpBB Notifier for macos
This script reads a configuration that defines phpBB forums and threads,
and watches them for new posts, refreshing on given interval.

When it finds new posts, it sends a native macos notification with the
thread title and post content.

It's meant for getting semi-realtime notifications about active but
somewhat sparsely updated discussions to help avoid having to check
manually all the time.

Use responsibly. Don't use this to hammer a server with too frequent
requests.

It uses Playwright to act as a browser to read threads, click for CLI
arguments, toml for parsing configuration and native osascript for
showing notifications.