Send notifications with terminal-notifier if able
When sending a notification, check if terminal-notifier
has been installed and if it has, use it for notifications.

It supports "click to open URL" mode that osascript doesn't,
which makes it a nicer option but I don't want to force users
to have that third party dependency installed so it defaults
to osascript.